Refactor(mentor): week7-add code spliting in App
