+++ src/App.js
@@ -1,7 +1,8 @@
 import { BrowserRouter, Routes, Route } from "react-router-dom";
-import SharedPage from "./pages/SharedPage";
-import FolderPage from "./pages/FolderPage";
-import ErrorPage from "./pages/ErrorPage";
+import { lazy } from "react";
+const SharedPage = lazy(() => import("./pages/SharedPage"));
+const FolderPage = lazy(() => import("./pages/FolderPage"));
+const ErrorPage = lazy(() => import("./pages/ErrorPage"));
 
 function App() {
   return (
